Add ID3v2 tags to the metadata table
ffmpeg doesn't convert ID3v2 tag names to generic metadata names, so
add the ID3v2 tag names to the table to pick them up.

This fixes scanning of MP3 files in various cases.

More in this post and its attachment:
<http://lists.mplayerhq.hu/pipermail/ffmpeg-devel/2009-September/076213.html>

Thanks to Raivo Hool for bringing up the issue and fix.