Use http_send_error to reply with an error
Allow to handle all error replies to add Access-Control-Allow-Origin headers. It
seems that firefox passes error responses without
Access-Control-Allow-Origin header to the javascript handler function but
nevertheless it is complaining about invalid CORS and it is not possible to get
the correct http return code. Therefore also add the header in error responses.