Use own error handling in httpd_send_error
It is not possible to set any header for the error response when using
evhttp_send_error. Calling evhttp_send_error results in cleaning all previously
set headers. Therefore it is not possible to set the Access-Control-Allow-Origin
CORS header in the error responses in conjunction with evhttp_send_error.

As a result an own error page must be generated and send back to the http
client. This allows to set any headers in httpd_send_error. Currently
httpd_send_error also cleans the previously set headers. Not sure if that is
really needed and should be better done in calling functions.

The new code in httpd_send_error is nearly the same as in evhttp_send_error
besides using internal api of libevent. Also it calls evhttp_send_error directly
if allow_origin config is not set or has an empty value.