Add a http handler method for CORS "preflight" requests
If some custom headers like "Authorization" are set in JavaScript the
XMLHttpRequest implementations will create a so called CORS preflight request
which is actually an OPTIONS http request with specific headers. For details
please take a look at http://www.html5rocks.com/en/tutorials/cors/ and
https://developer.mozilla.org/en-US/docs/Web/HTTP/Access_control_CORS.
This commit adds a handler method that creates a response to preflight
requests.