Fix avahi-daemon dependency in systemd init script.
'After:' only specifies when in the ordering that systemd will
attempt to start your unit file, however if one of the listed
services is disabled, it won't directly enable the disabled service.
'Requires:' should do exactly that: call the dependency that's needed
in order to start the service, regardless of its enablement state on
the system.