[tests] exploration to test MPD features
This is more of a proof-of-concept of how MPD could be tested.  Most
likely more is necessary to get some deterministic output, like
chrooting or masking in logs.

Whether or not that's worth the investment depends on whether a test
suite setup like this would be deemed useful and acceptable.

Signed-off-by: Fabian Groffen <grobian@gentoo.org>