[spotify] Fix misplaced librespot-c legacy mode selector
OwnTone was supposed to default to the old raw tcp Spotify protocol, which
would have left it unaffected by Spotify breaking the http response to the
metadata request. However, the mechanism for defaulting was misplaced. This
should put it in the right place and fix #1912.

Of course, the new protocol is still broken, but hopefully that will be fixed
on Spotify's side.