[scan] Read local file chapter metadata into mfi->chapters
Chapter metadata is an array of structs, one for each chapter. Saving that in
the normal way would bring much database complexity for a narrow use case, so
instead we just save it as json string in mfi->chapters. That's anyway the form
that we will be serving via the json API.

Should we need other forms in the future, we can use sqlite's json_each().