[dacp] Fix artwork for radio statios sometimes not updating
When "Ignore artwork provided by radio stations" is enabled, only the
queue_item->title would be used to create the bogus track id for Remote.
However, that is the name of the radio stations, so doesn't change.

This fixes the issue by including queue_item->album (which is actually
the track title when it's a radio station).

Fixes #1938